m00/ex01 - small update
Made the code a little more modular by making the input for the ADD
command a function and calling it for each piece of data.
Now we run this in a loop to make sure the user enters the info
correctly. Also we check to make sure the user didn't only enter
whitespace.
Other than this, a few comments explaining some stuff