big update to CPP_Module_03
Reworked a lot of the headers to remove the ifndef and replace with
pragma once.
also went through the cpp files and either cleaned things up or made
sure they tightly obeyed what what in the subject.
also for the main.cpp files, added some tests in a few cases.
Should be ready for push at this point.